Fixes UI race conditions, and adds a permanent http client
tcell 2.2.0 contains a bug with exiting that causes a deadlock in the UI
when run under Windows.  This was fixed in 2.2.1, however there is also a
race condition which is fixed in 2.3.1.  This also commits go.sum, which
should be present.  There is no guarantee that a dependency won't be
compromised in future, so this will lock everything to explicit commits.

This also splits how data is retrieved.  You're only meant to edit widgets
from the main tview loop.  This does the http access on the worker routine,
then pushes the updates in to the tview loop.  QueueUpdateDraw is sync
and will wait for the function to complete before resuming.

Finally we add a top level http.Client.  Creating a new client every time
will stop connection re-use.