Makes the active plot table columns sortable, and autorefresh logs
This is a relatively chunky change, which adds a new widget for
handling a sortable table, and moves the active plots table to
consume it.

There's also a change which renames plotTable to activePlotsTable,
and all associated data structures use this terminology as well.
This is intended to make things clearer by naming everything after
table titles (lastTable, aka Archived Plots in particular will
benefit from this)