Attempts to fix an issue where mouse input under PuTTy is not recognized
tcell.Screen exposes an EnableMouse with the option to specify which type of mouse events to
capture (click, drag, all).  Unfortunately putty doesn't currently support "all", which is the
default, so we need to specify either click or drag.

This problem is further compounded by tview.Application not exposing the underlying Screen
directly.  It does however use a Screen if we provide it; allowing us to create one, configure
the mouse, and then use it.  We don't call Application.EnableMouse because that will override
the setting we've used.